requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
httpx>=0.27
pytest>=8.3
pytest-asyncio>=0.24
supabase>=2.5
